♻[관리자]: FAQ 수정 모듈 분리
