+++ src/app/components/admin/faq/FAQGet.tsx
@@ -3,17 +3,17 @@
 import {
   ALERT_TEXT,
   BTN_TEXT,
-  FAQ_CATEGORY,
   INPUT_TEXT,
   TABLE_TEXT,
 } from '@/app/constants/admin';
 import { useCallback, useEffect, useState } from 'react';
-import { callDelete, callGet, callPut } from '@/app/utils/callApi';
+import { callDelete, callGet } from '@/app/utils/callApi';
 import Button from '../../common/Button';
 import Input from '../../common/Input';
 import Icons from '../../common/Icons';
 import { ArrowDownIcon } from '@/app/ui/iconPath';
 import FAQPost from './FAQPost';
+import FAQPut from './FAQPut';
 
 export default function FAQGet() {
   const [inputId, setInputId] = useState('');
@@ -68,47 +68,6 @@ export default function FAQGet() {
   const toggleContent = (faq_id: number) => {
     setSelectedId((prevId) => (prevId === faq_id ? null : faq_id));
     setEditingId(null); // 새로운 항목을 열 때 편집 모드 초기화
-  };
-
-  // FAQ 수정모드
-  const handleEdit = (
-    faq_id: number,
-    category: string,
-    question: string,
-    answer: string,
-  ) => {
-    setEditingId(faq_id);
-    setEditFAQ({
-      category,
-      question,
-      answer,
-    });
-  };
-
-  // 수정사항 저장
-  const handleEditSave = async () => {
-    if (
-      !editFAQ.category ||
-      editFAQ.category === '분류 선택' ||
-      !editFAQ.question ||
-      !editFAQ.answer
-    ) {
-      alert(ALERT_TEXT[0]);
-      return;
-    }
-    try {
-      await callPut(`/api/admin/faq/put/${selectedId}`, editFAQ);
-      alert(ALERT_TEXT[6]);
-      setEditFAQ({
-        category: '',
-        question: '',
-        answer: '',
-      });
-      setEditingId(null);
-      await handleGetFAQ(); // 저장 후 새로 불러오기
-    } catch (error) {
-      console.error(error);
-    }
   };
 
   // FAQ 삭제
@@ -174,91 +133,15 @@ export default function FAQGet() {
               </div>
 
               {selectedId === item.id && (
-                <>
-                  <div className="border-t-[1px] border-gray-4 mt-4 px-2 py-4 flex flex-col gap-4">
-                    {editingId === item.id ? (
-                      <>
-                        <select
-                          className="font-bold border p-2"
-                          value={editFAQ.category}
-                          onChange={(e) =>
-                            setEditFAQ((prev) => ({
-                              ...prev,
-                              category: e.target.value,
-                            }))
-                          }
-                        >
-                          {FAQ_CATEGORY.map((option) => (
-                            <option key={option} value={option}>
-                              {option}
-                            </option>
-                          ))}
-                        </select>
-                        <Input
-                          type="default"
-                          className="font-bold border p-2"
-                          textValue={editFAQ.question}
-                          onChange={(e) =>
-                            setEditFAQ((prev) => ({
-                              ...prev,
-                              question: e.target.value,
-                            }))
-                          }
-                        />
-                        <textarea
-                          className="border-2 py-2 px-4 font-bold min-h-44"
-                          value={editFAQ.answer}
-                          onChange={(e) =>
-                            setEditFAQ((prev) => ({
-                              ...prev,
-                              answer: e.target.value,
-                            }))
-                          }
-                        />
-                      </>
-                    ) : (
-                      <>
-                        <p className="border font-bold p-2">{item.question}</p>
-                        <div className="border-2 py-2 px-4">
-                          <p className="font-bold min-h-44">{item.answer}</p>
-                        </div>
-                      </>
-                    )}
-                  </div>
-                  <div className="flex gap-4 justify-end">
-                    <Button
-                      className={`${editingId === item.id ? 'bg-gray-2' : 'bg-red-1'} text-white font-bold max-w-fit px-8 py-2 self-end`}
-                      buttonText={
-                        editingId === item.id ? BTN_TEXT[5] : BTN_TEXT[1]
-                      }
-                      type="default"
-                      onClickHandler={
-                        editingId === item.id
-                          ? () => setEditingId(null)
-                          : handleDelete
-                      }
-                    />
-
-                    <Button
-                      className="bg-primary-3 text-white font-bold max-w-fit px-8 py-2 self-end"
-                      buttonText={
-                        editingId === item.id ? BTN_TEXT[6] : BTN_TEXT[3]
-                      }
-                      type="default"
-                      onClickHandler={
-                        editingId === item.id
-                          ? handleEditSave
-                          : () =>
-                              handleEdit(
-                                item.id,
-                                item.category,
-                                item.question,
-                                item.answer,
-                              )
-                      }
-                    />
-                  </div>
-                </>
+                <FAQPut
+                  editingId={editingId}
+                  selectedId={selectedId}
+                  editFAQ={editFAQ}
+                  setEditFAQ={setEditFAQ}
+                  setEditingId={setEditingId}
+                  handleGetFAQ={handleGetFAQ}
+                  item={item}
+                />
               )}
             </div>
           ))}
